+++ src/Sentry/Extensibility/DisabledHub.cs
@@ -181,6 +181,12 @@ public class DisabledHub : IHub, IDisposable
     }
 
     /// <summary>
+    /// No-Op
+    /// </summary>
+    public SentryId CaptureCheckIn(string monitorSlug, CheckInStatus status, SentryId? sentryId = null)
+        => SentryId.Empty;
+
+    /// <summary>
     /// No-Op.
     /// </summary>
     public Task FlushAsync(TimeSpan timeout) => Task.CompletedTask;