+++ src/Sentry/Extensibility/HubAdapter.cs
@@ -268,6 +268,12 @@ public sealed class HubAdapter : IHub
         => SentrySdk.CaptureSession(sessionUpdate);
 
     /// <summary>
+    /// Forwards the call to <see cref="SentrySdk"/>.
+    /// </summary>
+    public SentryId CaptureCheckIn(string monitorSlug, CheckInStatus status, SentryId? sentryId = null)
+        => SentrySdk.CaptureCheckIn(monitorSlug, status, sentryId);
+
+    /// <summary>
     /// Forwards the call to <see cref="SentrySdk"/>
     /// </summary>
     [DebuggerStepThrough]